Rename sector config update file
