+++ scripts/lib/validation/config/sector_eurelectric_config.py
